libeufi: Fix some gtk-doc annotations
This hash table returned has pointer array elements. Also, these should
be marked (transfer container) not (transfer full), because freeing the
GHashTable container already properly frees the elements in it. With
(transfer full) the values in the hash table are freed too many times,
which causes Python to hit a memory error.

https://phabricator.endlessm.com/T21756