tests: Check number of received warning messages
Now that we emit warnings rather than debug messages for various
non-terminal problems with the autoinstall files, the tests need to
expect those warnings to be emitted. Change them to do so.

Ideally, we’d have a test fixture which reset the warning count to 0 at
the start of the test, and asserted it was zero at the end of the test;
but adding that is a bit too invasive for such a small branch. We don’t
expect this code to change much, so we can get by with an assertion at
the end of main(). That just means any (unlikely) failure will be a bit
harder to pin down instantly.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T22054