libeos-updater-util: Pull deps from remotes with no collection ID
If we don’t have a collection ID in the local configuration for a
remote, we still want to use it to resolve dependencies — otherwise we
won’t resolve any dependencies from any remotes at the moment (since
almost all deployments don’t have collection IDs in their configurations
yet).

The original argument against allowing this
(https://github.com/endlessm/eos-updater/pull/189#issuecomment-382601291),
was that we wouldn’t know what collection ID to use for pulling the
dependency refs (and that we couldn’t use the collection ID from the app
ref, since the dependencies might be coming from a different
collection). While the latter is true, the former is not: if collection
IDs are not configured locally, we would not be pulling with them
anyway (we’d be pulling using the remote name). If they are configured,
we can use them. To this end, the code handles NULL
EuuFlatpakLocationRef:collection-id properties throughout.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T22054