eos-updater: Allow upgrades even if the current commit is missing
This is the final commit in a short series to add support for upgrading
even if the .commit object for the currently deployed commit has
(somehow) been deleted from the client’s repository.

This includes unit tests for eos-updater and
eos-updater-flatpak-installer to ensure they both work if the commit has
been deleted.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T22805

(eos3.3 backport: Dropped an irrelevant chunk for handling version
metadata.)