tests: Add missing g_assert_no_error() calls
There was a systematic lack of g_assert_no_error() calls after certain
utility calls in the tests, which could have resulted in the setup for
the test silently failing. Since setup functions like
eos_test_setup_flatpak_repo_simple() are not atomic (failure can leave
the repo half-configured), such failures might not be noticed later on
in the tests.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T22873