eos-updater: Loosen an invariant on autoinstall entry collection IDs
We previously had an invariant that after loading in an autoinstall
list, the collection ID and the remote name had to be set in each entry
(possibly by looking one up from the other). This was fine for apps
listed in the autoinstall list; but not for their dependencies, for
which collection ID information is typically not available in the app
metadata.

Loosen that invariant so that the collection ID isn’t required (but if
it is set, it’s validated). See the comment in the code for more
details.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T22054