libeos-updater-util: Resolve remote name for upgrade/uninstall ops
The upgrade and uninstall entries in an autoinstall list don’t specify a
remote name to use: we need to look it up from the installed origin of
the app in question. We weren’t doing that, though, meaning that ‘none’
was being used as a remote name. This caused lookups of dependencies
(such as runtimes) to quietly fail.

Fix this by setting the remote name to NULL in those cases, and
bypassing the lookup of related refs if it’s NULL.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T22873