eos-updater: Downgrade a warning to a message depending on the error
If we fail to poll for P2P updates due to it not being supported by the
configuration (i.e. no collection ID is configured), emit a message
rather than a warning. We should only emit a warning if something which
is expected to work, is not working. If the P2P configuration is
missing, then it’s not expected to work.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T22873