tests: Add a FIXME about blocking subprocesses by not reading from pipes
It turns out we do not asynchronously read from the stdout/stderr pipes
from subprocesses spawned by the tests: we wait until the subprocess has
finished before reading from the pipes. This means all the subprocess’
output builds up in the OS pipe buffer. If a subprocess produces more
output than can be stored in the pipe buffer, it will block
indefinitely. This can happen if the subprocesses are run with
G_MESSAGES_DEBUG=all.

Fixing this is non-trivial and best done by rewriting the test spawn
code to use GSubprocess and asynchronously wait on subprocesses while
reading their stdout/stderr. It’s too big a task to do now, so leave a
FIXME in place and defer to https://phabricator.endlessm.com/T18883.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T22873