UNTESTED eos-updater: Do a prune before fetching updates
Prune the repository before fetching updates from the server, to
increase the chance that we have enough disk space to complete the
fetch.

This seems like a reasonable place to put a prune operation, since we
already hold the repository lock, and are already running as a
background process on a regular schedule.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T22763