tests: Allow ignoring certain files when doing `rm -rf`
eos_updater_remove_recursive() is essentially `rm -rf`. For an upcoming
new test, it would be useful to limit the removals to files which match
a predicate. Add a predicate callback to eos_updater_remove_recursive().

Also take the opportunity to remove a pointless wrapper function,
rm_rf().

This introduces no functional changes for the way we currently call
eos_updater_remove_recursive().

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T22805