poll: Fix an assertion failure on the fallback code path
In metadata_fetch_internal(), there are two code paths for getting
metadata, a "new" one that supports pulling from peer sources, and a
fallback one that only fetches from the Internet. On the fallback code
path, we are passing to run_fetchers() only one fetcher but the complete
configured list of sources. Since the fallback path is supposed to only
be used for fetching from the Internet, we should only pass that source
to run_fetchers(). This commit does that, which avoids an assertion
failure from the line "g_return_val_if_fail (fetchers->len ==
sources->len, NULL);".

https://phabricator.endlessm.com/T22192