poll: Fix following redirects with collection refs
The code that followed redirect commits when collection IDs are in use
is broken. It basically ignores the newly redirected ref and uses the
booted one instead. So this commit fixes that bug, so that redirects
work when collection IDs are enabled.

Also, add a unit test for following eol-rebase redirects, which would
have caught this bug.

https://phabricator.endlessm.com/T22192