eos-updater: Fix a regression of the offline support
When check_for_update_following_checkpoint_if_allowed() was added to
metadata_fetch_new() in 116ce762d it broke the use case of doing offline
OS updates, because it indirectly uses fetch_latest_commit() which
assumes an Internet connection in its use of
ostree_repo_pull_with_options(). The effect is that offline OS updates
currently error out with "Error resolving origin.ostree.endlessm.com".

So this commit changes fetch_latest_commit() so that it uses
ostree_repo_find_remotes_async() when appropriate (basically, when a
collection ID is configured) so that USB and LAN sources are checked for
updates (if the eos-updater configuration allows it). This commit also
makes changes so that fetch_latest_commit() has the collection-ref and
OstreeRepoFinder instances it needs to do the P2P operations.

The way it's written right now there's redundancy between what
check_for_update_following_checkpoint_if_allowed() does and what
metadata_fetch_new() does, but that will be addressed in a follow-up
commit.

https://phabricator.endlessm.com/T22192