apply: Remove useless message when applying an update
It’s not useful to know that the boot version didn’t change. The boot
version is an integer used to disambiguate between two deployments of
the same OS and commit checksum, so is almost always 0 (especially on
EOS systems, where we bail a lot earlier in the process if we would have
ended up redeploying the same commit checksum).

Remove this pointless message and the support infrastructure for it.
Some more useful messages will be added in a following commit.

Signed-off-by: Philip Withnall <withnall@endlessm.com>