eos-updater-flatpak-installer.service: loosen sandboxing for bwrap
The systemd service sandboxing for the boot-time Flatpak installer prevents the
Flatpak triggers from working correctly, because they are executed via bwrap.
One easily visible consequence is that the mime cache in the flatpak
export/share/applications folder is not updated, meaning that documents are not
associated with a newly-installed LibreOffice until an unrelated (and
unconstrained) Flatpak operation happens to re-run the triggers. Add the (sadly
quite broad) permissions necessary to allow bwrap to work correctly.

https://phabricator.endlessm.com/T22900