flatpak: Don't search broken remotes for dependencies
Instead of erroring out completely, blocking updates, just log
a message indicating that the remote is broken and don't search
it for things like runtimes or related refs. We require things
like summary files to be present on remotes and this will be broken
the moment a user installs a flatpak bundle since those don't have
summaries generated on their remote configuration.

https://phabricator.endlessm.com/T23347