build: Remove check for ostree experimental API
As of ostree 2018.6, the P2P API is public and since we have 2018.6 in
Endless now we no longer need to check for the "experimental" feature
flag. Also, bump the required versions of flatpak and ostree so we can
be sure the P2P API is public in both. See:
https://github.com/ostreedev/ostree/commit/8fbf19c9f54ba34ae7ebbc540b08566b8dd18848

https://phabricator.endlessm.com/T23138