test-common: Use the flatpak --assumeyes option
Recent verisons of flatpak prompt the user for confirmation during
install/update/uninstall operations. So when the eos-updater unit tests
try to install the test runtime, there's an "Is this ok? [y/n]:"
prompt, and no is assumed since there's no user input, which eventually
leads to a unit test failure.

So this commit adds the --assumeyes option when the tests call out to
`flatpak install` or `flatpak uninstall`, which is meant for the purpose
of scripting and answers yes to questions (or the default option if it's
multiple choice).

https://phabricator.endlessm.com/T23138