poll: Comment check_for_update_following_checkpoint_commits
There's a bunch of different refspecs, refs and collection_refs
being thrown around here, which is confusing for the reader. Document
exactly what variables serve what purpose.

(This commit is the cherry-pick for the eos3.3 branch. The only
difference is that on this branch the UpdateRefInfo struct has no
"results" member; the original commit moved a statement modifying that
member down a few lines.)