eupv: Exclude apps that are installed then uninstalled
Currently eos-updater-prepare-volume looks at the flatpak autoinstall
lists and simply includes all the ones that are set to be installed or
updated and none of the ones set to be uninstalled. But some apps such
as com.endlessm.CompanionAppService//eos3 have an install action and a
later uninstall action. So currently eos-updater-prepare-volume tries to
include that app on a USB, which fails since it's not installed. So this
commit fixes e-u-p-v so that for each uninstall action, it removes any
previous install/update actions in the set with the same ID.

https://phabricator.endlessm.com/T21756
(cherry picked from commit a003a861381547ced4f203ce50cab99c1d11a94f)
Signed-off-by: Robert McQueen <rob@endlessm.com>