eupv: Flatten flatpak actions using library API
The set of flatpak actions that can be configured for eos-updater are
each "install", "update", or "uninstall" operations, but a couple things
affect what the final set of actions applied will be. One is that there
are directories in /var and /etc whose configuration files override the
ones in /usr. Additionally, each action has an associated serial number
and later uninstall actions cancel out a previous install action for the
same ID.

The way e-u-p-v currently deals with this is to subtract the uninstall
actions from the install actions, but this is imperfect because for
example an app could conceivably be installed, later uninstalled, and
then installed again. In that case it should be included on the update
USB but isn't by the current code.

So this patch uses library API to flatten the set of actions so that
there's only one for each app, which is more robust. It's still
imperfect because if there are overrides configured on the computer
creating the USB or on the computer using the USB to update, that would
mean the set of apps on the update USB is not exactly the set needed for
the update. But I don't think we're likely to hit those cases in actual
use in the foreseeable future.

See previous discussion of this topic here:
https://github.com/endlessm/eos-updater/pull/188#discussion_r188607962

https://phabricator.endlessm.com/T21756