libeuu: Add some convenience API
This commits adds a library function,
euu_flattened_flatpak_ref_actions_from_paths(), that combines
euu_flatpak_ref_actions_from_paths() and
euu_flatten_flatpak_ref_actions_table(). In Python if you call the first
you get an object of type `dict`, and if you try to pass that to the
second function that expects a GHashTable you get an error. So this API
will allow eos-updater-prepare-volume to get what it needs working
within the apparent limits of the pygobject bindings.

https://phabricator.endlessm.com/T21756