eos-updater: set IO class to idle
Our benchmarks show this significantly reduces the interactivity impact of
ongoing ostree operations while the user is continuing other tasks on the
system. The effect is very pronounced with the default CFQ scheduler, and in
combination with BFQ, using the idle class improves the worst case to nearly
the same as an unloaded system.

https://phabricator.endlessm.com/T22479