eupv: Run "chmod -R 777" on the ".ostree" directory
This commit has eos-updater-prepare-volume run "chmod -R 777" on the
".ostree" directory in a destination USB drive. Since
eos-updater-prepare-volume runs as root but "flatpak create-usb" runs as
the user, we need the repo on the USB drive to be writable by both. And
generally if you're putting something on a USB, you want anyone in
possession of the USB to be able to read and write it; ext4 permissions
are a bit unnecessary in this context.

For this not to be a destructive operation the repo mode used can't
store permission information directly in the file attributes. I've just
submitted PRs to Flatpak and OSTree to make that the case.

https://phabricator.endlessm.com/T23391