eupv: Copy the booted commit to the USB
Currently eos-updater-prepare-volume calls `ostree create-usb` and
passes it the deployed ref, like os/eos/amd64/eos3a. Then `ostree
create-usb` might fail if the commit pointed to by that ref is partial,
which is what happens when eos-updater fetches the latest available
commit in its Poll stage. This effectively means that any user not on
the latest version of the OS can't copy the OS to a USB after
eos-updater first checks for updates, which happens automatically in the
background.

So this commit uses the newly added --commit option of the create-usb
command, so that we're copying the booted commit (which must be
complete) rather than the potentially partial commit pointed to by the
ref.

https://phabricator.endlessm.com/T23851