poll: Better determine if an update was found
Currently in metadata_fetch_new() we check if a commit was found that
appears to be an update, and assume that if so the array of
OstreeRepoFinderResults is non-empty. However this is not a safe
assumption. If for example you have a newer OS commit in your local repo
(because you used eos-rollback or fetched it in a previous run of
eos-updater), this commit is read by parse_latest_commit() and we then
move on to the Fetch() stage where libostree hits the assertion failure:

ostree_repo_pull_from_remotes_async: assertion 'results != NULL &&
results[0] != NULL' failed

So the fix is to check if an OstreeRepoFinderResult was actually found
in metadata_fetch_new() when deciding whether an update was found. This
is especially important now that we're checking offline and online
sources separately, because that means we need to distinguish "an update
being available" from "an update being available from one of the sources
being checked currently".

https://phabricator.endlessm.com/T23972