eos-updater-prepare-volume: Fix collection ID formatting
Commit 784b55f04d8e4ff8d0e38f63283b0e7d19da316a changed how collection
IDs were handled, and started extracting them from the GVariants in
commit metadata.

Using `str(variant)` in Python results in a call to `g_variant_print()`,
which formats the variant as a string — for variants of type `s`, this
means adding single quotes around it. So our collection ID ended up
being `"'com.endlessm.SomeId'"`.

This caused problems when used further down the chain.

Fix that by using `variant.get_string()` instead.

This problem wasn’t caught when the commit above was originally made, as
the OBS build of eos-updater was broken, and hence the installed tests
(which caught this problem) weren’t being run.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T26056