test-common: Adjust parsing of `flatpak list` output
Since flatpak improved its console UI, the format of `flatpak list` has
changed; adjust our parsing of it, and pass the `--columns` option to
ensure we get the fields we want. We don’t actually need to update the
parsing glob and code, but do need to update the documentation around
it, and the way we call `flatpak list`.

This bumps our flatpak dependency to 1.1.2 (which we probably already
silently depended on).

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T26056