tests: Disable parental controls checks for libeufi tests
Use the `FLATPAK_SYSTEM_HELPER_ON_SESSION` environment variable to get
libflatpak to skip its parental controls checks when installing apps and
runtimes.

While we could set parental controls metadata for the fake apps used in
the tests, that’s not the problem — the problem is that on OBS, there is
no D-Bus system bus running, so libflatpak cannot talk to the
accountsservice to get the user’s parental controls policy. This makes
it assume that the parental controls policy is maximally restrictive,
and hence it prevents installation of the app.

`FLATPAK_SYSTEM_HELPER_ON_SESSION` causes libflatpak to use the session
bus rather than the system bus for talking to the system helper, and
coincidentally is used as an escape hatch to get it to bypass parental
controls checks. This is a hack. See
https://phabricator.endlessm.com/T25340 for the proper fix.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T26056