eupv: Use ref(s) from commit metadata
Currently eos-updater-prepare-volume checks the deployment's origin
refspec to decide what ref to pass to "ostree create-usb" to copy the OS
onto a USB drive. This is not always correct; for example images built
from development OSTrees will have the production ref (e.g.
os/eos/amd64/eos3a) set as the origin so that it will be used for
updates, but the booted commit will have a different associated ref
(e.g. os/eos/amd64/master). In that case the "ostree create-usb" command
will fail. So this commit changes e-u-p-v so that it checks the commit
metadata to find the list of associated refs and copies each of those. I
think it makes more sense to copy all of them than to try to determine
which will be used on the receiving computer, and the additional disk
space usage is trivial.

In other words, this commit should ensure that e-u-p-v works on
development images, not just production ones (where the booted and
update refs often match).

https://phabricator.endlessm.com/T24034