eupv: Add a --preserve-permissions argument
Currently eos-updater-prepare-volume changes the permissions on the
drive being written to so that any user can read the contents of its
toplevel directory (so they have access to the .ostree directory). This
is a bit rude and may be unexpected, so add an argument that allows this
behavior to be disabled. This is intended for manual use; GNOME Software
won't be able to use it.

https://phabricator.endlessm.com/T25940