test-common: Define XDG_CACHE_HOME for e-u-p-v
Now that we're using G_TEST_OPTION_ISOLATE_DIRS, GLib is setting several
environment variables (HOME, XDG_CACHE_HOME, etc.) to "/dev/null" since
the tests are expected to use the helper functions instead such as
g_get_user_cache_dir (). This works fine for the tests themselves but
not subprocesses such as when we're spawning
"eos-updater-prepare-volume". So we need to explicitly pass on
XDG_CACHE_HOME since flatpak uses it in
flatpak_get_user_cache_dir_location() and otherwise we get the error
message "error: Not a directory" during test-update-from-volume. I'm not
sure why this failure happens locally but not on Jenkins/OBS.

Even with this patch, the test "test-update-from-volume" still fails but
with a different error; see the ticket for discussion of that.

https://phabricator.endlessm.com/T26806