tests: Add service to start flatpak-system-helper
In order for test-update-from-volume to work we need an instance of the
flatpak-system-helper on the session bus, since we set
FLATPAK_SYSTEM_HELPER_ON_SESSION in the environment. So add a service
file based on the one Flatpak uses for its own tests, and add it to the
set of directories GTestDBus searches, to avoid the error "The name
org.freedesktop.Flatpak.SystemHelper was not provided by any .service
files".

https://phabricator.endlessm.com/T26806