eos-updater-flatpak-installer: skip parental controls checks
Without this, deploying Flatpaks after a reboot fails with the following
in the journal (line-wrapped for legibility):

   Attempting to install eos-apps:app/com.hack_computer.Clubhouse/x86_64/eos3
   Failed to install eos-apps:app/com.hack_computer.Clubhouse/x86_64/eos3:
     Could not connect: No such file or directory
   eos-updater-flatpak-installer: Couldn’t apply some flatpak update actions for this boot:
     Could not connect: No such file or directory

This problem does not occur if I trigger this service once the system is
fully booted.

I believe this is because deploying a Flatpak now involves a parental
controls check, which involves talking to accountsservice on the system
bus, neither of which are running at the point where this service runs.
Work around this by using the same workaround we used in the image
builder to skip the parental controls check.

https://phabricator.endlessm.com/T28530