libeos-updater-util: Update flatpak deps in correct order
Currently when we're preparing the list of flatpaks to install which are
dependencies of flatpaks configured for auto-installation as part of an
OS update, we use hash tables in a couple places, which means they're
not guaranteed to be installed in any particular order. This is a
problem because libreoffice depends on org.freedesktop.Platform//19.08,
which has as one of its extensions org.freedesktop.Platform.openh264,
and the openh264 extension uses the "extra-data" feature where
downloaded software is unpacked into the runtime at deploy time. This is
only possible when the runtime is already installed, so when eos-updater
attempts to install the extension first it hits this error:

Fetch: failed to pull necessary new flatpaks for update: While trying to
apply extra data: runtime/org.freedesktop.Platform/x86_64/19.08 not
installed

Fix the problem by changing a data structure to use a pointer array
instead of a hash table when the list of dependencies is being
enumerated (since the runtime is added before its extensions). Also,
change squash_ref_actions_ptr_array() so that it preserves the order of
the pointer array while compressing the actions. I think the performance
impact of these changes will be negligible; in practice we're talking
about less than 30 flatpaks.

(This commit is for the master branch; the only difference between it
and the commit on eos3.3 is the removal of a (GDestroyNotify) cast in
euu_flatpak_related_refs_for_remote_free().)

https://phabricator.endlessm.com/T28626