poll: Use ostree_commit_get_object_sizes if available
OSTree upstream recently gained an API to return the parsed entries in
the `ostree.sizes` metadata[1]. Split out a helper function and use it
if it's available to sum the various commit size properies. There is one
slight behavior change here - the updater only continues on error if the
error is `G_IO_ERROR_NOT_FOUND`. This is what both APIs return when
`ostree.sizes` does not exist as well as what's now set when neither
APIs are available.

1. https://github.com/ostreedev/ostree/pull/1957

https://phabricator.endlessm.com/T18171