eos-updater: Remove the Mogwai schedule entry after a successful download
The schedule entry was successfully removed on all the error paths, but
accidentally not the success path. That meant it hung around until
either `eos-updater` or `mogwai-scheduled` were restarted. Since
Mogwai’s default policy is to only allow one download in parallel, it
blocked all other downloads if the connection was metered.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T29311