libeos-updater-util: Rename some headers to avoid clashes
If the current source directory (libeos-updater-util) gets into the
include path, these headers get included instead of the system-installed
flatpak, OSTree, etc. headers, which causes a compilation failure.

In particular, this is a problem in the upcoming Meson port.

It doesn’t seem possible to remove a directory from the include path, so
rename the files instead.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T22069