eos-autoupdater: Remove static assertion
The upcoming port to Meson breaks this and I can’t figure out why:
`G_LOG_DOMAIN` is definitely defined, so it’s unclear why it can no
longer be used in a static assertion.

Both autotools and Meson define it on the compiler command line using
`-D`.

The check seems slightly unnecessary, since the code will fail to
compile lower down if `G_LOG_DOMAIN` is not defined to a string — so
rather than spend time debugging the problem, just remove the check.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T22069