Re-implement flatpak dependency logic using transactions
This commit changes the way we enumerate dependencies (runtimes and
related refs) of flatpak ref actions (that is, flatpaks that are set to
be installed/updated/uninstalled during an OS upgrade) so that we rely
on the FlatpakTransaction API rather than re-implementing large parts of
it.

This commit also passes the value of g_get_user_cache_dir() to the
subprocess' XDG_CACHE_HOME env var, which would otherwise be set to
/dev/null due to G_TEST_OPTION_ISOLATE_DIRS being set. Flatpak needs the
cache dir for flatpak_dir_create_child_repo().

Subsequent commits will make use of the FlatpakTransaction API for the
execution of install/update/uninstall actions.

https://phabricator.endlessm.com/T23150