tests: Add more debug output and refactor
This commit changes the tests that expect an attempted update to fail so
they share more code. This also changes them so that we get log output
from the failures as we do on the success code path
(etc_update_client()). We also now consistently check the exit status
for failure, which previously we were doing sporadically in some of the
tests.