tests: Update for new runtime resolution behavior
When these tests were written, eos-updater resolved flatpak runtime
dependencies. Now we let Flatpak do it. As discussed in this PR[1], the
behavior decided on upstream is for an app's origin to be searched
before other remotes with the same priority, so update the unit tests
here to make it more clear what we're testing.

[1] https://github.com/flatpak/flatpak/pull/3751