libeos-updater-util: Fix FlatpakTransaction operation-error handling
This fixes a regression from T23150, which ported eos-updater to use
`FlatpakTransaction`. It turns out that if you don’t connect to
`FlatpakTransaction::operation-error`, any error is fatal, including
errors such as `FLATPAK_ERROR_SKIPPED`, or non-fatal errors for
nice-to-have related refs.

It also turns out that even if a transaction contains only one
operation, the error returned by `flatpak_transaction_run()` on failure
is always a generic one. `operation-error` is needed to get the detailed
error message. Fixing this means we can revert the changes to the error
codes in the unit tests from commit 200fb988da.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T25765